Update & Use utilities 1.2
