+++ plugin/src/main/java/net/codingarea/challenges/plugin/challenges/implementation/challenge/ForceBiomeChallenge.java
@@ -11,7 +11,7 @@ import net.codingarea.challenges.plugin.management.server.scoreboard.ChallengeBo
 import net.codingarea.challenges.plugin.utils.item.ItemBuilder;
 import net.codingarea.challenges.plugin.utils.misc.NameHelper;
 import net.codingarea.challenges.plugin.utils.misc.RandomizeUtils;
-import net.codingarea.challenges.plugin.utils.misc.StringUtils;
+import net.anweisen.utilities.commons.misc.StringUtils;
 import org.bukkit.Material;
 import org.bukkit.block.Biome;
 import org.bukkit.boss.BarColor;
